refactor: refactor function generateProduct
